refactor: remove BaseLiqMapCard abstraction
