+++ src/lib/comps/charts/chartUtils.ts
@@ -319,9 +319,12 @@ export async function fetchLiqMapDataMerged(
 	};
 }
 
-export async function getSupportedLiqMapInstrumentOptions(): Promise<
-	{ label: string; value: InstrumentInfo }[]
-> {
+export type InstrumentOption = {
+	label: string;
+	value: InstrumentInfo;
+};
+
+export async function getSupportedLiqMapInstrumentOptions(): Promise<InstrumentOption[]> {
 	const coinstatsSelectedCoin = get(coinstats_selected_coin);
 
 	if (!coinstatsSelectedCoin) {
